pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fast-api-db-template"
version = "1.0.0"
description = "FastAPI template with database integration"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "alembic",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "asyncpg",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
